Stabilize FAI-655 quality and APIGen checks
